Update .info & Readme files for distribution.
I have added an extension page here
https://civicrm.org/extensions/export-ui

and with these changes we should be able to publish a downloadable release.

I would like to see us start to plan towards having this ship-enabled & then being moved into
core but we should probably get a 'proper' release out there as our first step.

We have hit no problems in production. I'm not 100% sure of the minimum version - since we are using on
5.18 and it is the current stable as well. I would expect us to increment the version
if we make any improvements & don't see any benefit in figuring out the technical minimum